+++ tests/test_queries.py
@@ -95,6 +95,42 @@ class TestMongoQueries(SingleCollectionTest):
         self.assertEqual(total, 150)
 
     @defer.inlineCallbacks
+    def test_FindWithCursorBatchsize(self):
+        yield self.coll.insert([{'v': i} for i in range(140)], safe=True)
+
+        docs, d = yield self.coll.find_with_cursor(batch_size=50)
+        lengths = [] 
+        while docs:
+            lengths.append(len(docs))
+            docs, d = yield d
+        self.assertEqual(lengths, [50,50,40])
+
+    @defer.inlineCallbacks
+    def test_FindWithCursorBatchsizeLimit(self):
+        yield self.coll.insert([{'v': i} for i in range(140)], safe=True)
+
+        docs, d = yield self.coll.find_with_cursor(batch_size=50,limit=10)
+        lengths = [] 
+        while docs:
+            lengths.append(len(docs))
+            docs, d = yield d
+        self.assertEqual(lengths, [10])
+
+    @defer.inlineCallbacks
+    def test_FindWithCursorZeroBatchsize(self):
+        yield self.coll.insert([{'v': i} for i in range(140)], safe=True)
+
+        docs, d = yield self.coll.find_with_cursor(batch_size=0)
+        lengths = [] 
+        while docs:
+            lengths.append(len(docs))
+            docs, d = yield d
+        print(lengths)
+        self.assertEqual(lengths, [101,39])
+
+
+
+    @defer.inlineCallbacks
     def test_LargeData(self):
         yield self.coll.insert([{'v': ' '*(2**19)} for _ in range(4)], safe=True)
         res = yield self.coll.find()