+++ txmongo/collection.py
@@ -367,7 +367,7 @@ class Collection(object):
 
     @timeout
     def find_with_cursor(self, *args, **kwargs):
-        """find_with_cursor(filter=None, projection=None, skip=0, limit=0, sort=None, **kwargs)
+        """find_with_cursor(filter=None, projection=None, skip=0, limit=0, sort=None, batch_size=0, **kwargs)
 
         Find documents in a collection and return them in one batch at a time.
 
@@ -389,7 +389,8 @@ class Collection(object):
         new_kwargs = self._find_args_compat(*args, **kwargs)
         return self.__real_find_with_cursor(**new_kwargs)
 
-    def __real_find_with_cursor(self, filter=None, projection=None, skip=0, limit=0, sort=None, **kwargs):
+    def __real_find_with_cursor(self, filter=None, projection=None, skip=0, limit=0, sort=None, batch_size=0,**kwargs):
+        
         if filter is None:
             filter = SON()
 
@@ -401,6 +402,8 @@ class Collection(object):
             raise TypeError("TxMongo: skip must be an instance of int.")
         if not isinstance(limit, int):
             raise TypeError("TxMongo: limit must be an instance of int.")
+        if not isinstance(batch_size, int):
+            raise TypeError("TxMongo: batch_size must be an instance of int.")
 
         projection = self._normalize_fields_projection(projection)
 
@@ -414,8 +417,15 @@ class Collection(object):
 
             check_deadline(kwargs.pop("_deadline", None))
 
+            if batch_size and limit:
+                n_to_return = min(batch_size,limit)
+            elif batch_size:
+                n_to_return = batch_size
+            else:
+                n_to_return = limit
+
             query = Query(flags=flags, collection=str(self),
-                          n_to_skip=skip, n_to_return=limit,
+                          n_to_skip=skip, n_to_return=n_to_return,
                           query=filter, fields=projection)
 
             deferred_query = protocol.send_QUERY(query)
@@ -427,8 +437,10 @@ class Collection(object):
         # reference between closure and function object which will add unnecessary
         # work for GC.
         def after_reply(reply, protocol, this_func, fetched=0):
+
             documents = reply.documents
             docs_count = len(documents)
+
             if limit > 0:
                 docs_count = min(docs_count, limit - fetched)
             fetched += docs_count
@@ -439,8 +451,13 @@ class Collection(object):
             out = [document.decode(codec_options=options) for document in documents[:docs_count]]
 
             if reply.cursor_id:
+                # please note that this will not be the case if batch_size = 1
+                # it is documented (parameter numberToReturn for OP_QUERY)
+                # https://docs.mongodb.com/manual/reference/mongodb-wire-protocol/#wire-op-query 
                 if limit == 0:
                     to_fetch = 0  # no limit
+                    if batch_size:
+                      to_fetch = batch_size
                 elif limit < 0:
                     # We won't actually get here because MongoDB won't
                     # create cursor when limit < 0
@@ -449,6 +466,8 @@ class Collection(object):
                     to_fetch = limit - fetched
                     if to_fetch <= 0:
                         to_fetch = None  # close cursor
+                    elif batch_size:
+                        to_fetch = min(batch_size,to_fetch)
 
                 if to_fetch is None:
                     protocol.send_KILL_CURSORS(KillCursors(cursors=[reply.cursor_id]))